refactor: extract credential lifecycle mixin from coordinator core
